Module tidb_query_vec_executors::interface

Batch executor common structures.

Structs

BatchExecuteResult
Data to be flowed between parent and child executors' single next_batch() invocation.

ExecuteStats
Execution statistics to be flowed between parent and child executors at once during collect_exec_stats() invocation.

WithSummaryCollector
Combines an ExecSummaryCollector with another type. This inner type T typically Executor/BatchExecutor, such that WithSummaryCollector<C, T> would implement the same trait and collects the statistics into C.

Traits

BatchExecutor
The interface for pull-based executors. It is similar to the Volcano Iterator model, but pulls data in batch and stores data by column.

ExecSummaryCollector
A trait for all execution summary collectors.