Module tidb_query_vec_executors::util::scan_executor

Structs

ScanExecutor
A shared executor implementation for both table scan and index scan. Implementation differences between table scan and index scan are further given via ScanExecutorImpl.

ScanExecutorOptions

Traits

ScanExecutorImpl
Common interfaces for table scan and index scan implementations.

Functions

check_columns_info_supported
Checks whether the given columns info are supported.

field_type_from_column_info
Extracts FieldType from ColumnInfo.